Net: Add temporary solution for @nogc toString
Temporary, needs fixing. Currently, the SockAddrIPv4 and SockAddrIPv6
return different *types* (FixedString length is different). This means
we cannot unite them.

We need a solution that will work over SockAddr, as well as something
that will allow creating the whole string (including the port) without
multiple copying of the string.