Fix build with -dip25 and prevent regressions
Functions that return by `ref` their formal parameters (or the `this`
implicit parameter) need to have the `return` attribute, so the compiler
can catch cases where the return value outlives the function argument it
was derived from. This attribute is inferred (just like e.g. `nothrow`
or `pure`) for function templates (or functions defined inside
templates), nested functions and `auto`-returning functions.

The `-dip25` flag enforces this static analysis and it  will become the
default in one the future releases. This pull request enables this flag
and fixes the compilation errors.